- Create and add the new Transactions, Merchants and Settings icons and use them in the Balances toolbar. - Remove the transactions section from the Balances fragment. - Create a new TransactionsFragment which hosts the list of trasactions. - Add the transactions destination to the navigation graph so that the Navigation component automatically handles the transition when the transactions icon is selected in Balances.
