+++ app/src/main/res/layout/fragment_balances.xml
@@ -35,20 +35,6 @@
             tools:listitem="@layout/item_balance"
             tools:itemCount="3"/>
 
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="@dimen/spacing_same_topic"
-            android:text="@string/title_transactions"
-            android:textAppearance="@style/TextAppearance.Bitsy.Subtitle1"/>
-
-        <androidx.recyclerview.widget.RecyclerView
-            android:id="@+id/rvTransactions"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            tools:listitem="@layout/item_transaction"
-            tools:itemCount="6"/>
-
     </LinearLayout>
 
 </ScrollView>
